requirements.txt
selenium==4.10.0
webdriver-manager==4.0.0
Pillow==10.0.0
numpy==1.24.4
opencv-python==4.8.0.74
matplotlib==3.7.2
seaborn==0.12.2
fastapi==0.101.0
 uvicorn==0.23.2